# packages/api-zod/src/generated/types/alertWithMother.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MatriWatch maternal health monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface AlertWithMother {
  id: string;
  mother_id: string;
  /** @nullable */
  mother_name?: string | null;
  clinic_id: string;
  alert_type: string;
  message: string;
  is_read: boolean;
  created_at: string;
}